Change RNG type to unsigned long

+++ lib/rng.c
@@ -9,7 +9,7 @@ void rl_rng_stdlib_init()
     srand(time(0));
 }
 
-int rl_rng_stdlib_generate(int min, int max)
+unsigned long rl_rng_stdlib_generate(unsigned long min, unsigned long max)
 {
     // TODO assert/unsigned?
     if (min < 0 || max < 0 || (min == 0 && max == 0))
@@ -32,12 +32,12 @@ void rl_rng_twister_free()
     if (rl_twister_state) mtwist_free(rl_twister_state);
 }
 
-int rl_rng_twister_generate(int min, int max)
+unsigned long rl_rng_twister_generate(unsigned long min, unsigned long max)
 {
     // TODO assert/unsigned?
     if (min < 0 || max < 0 || (min == 0 && max == 0))
         return 0;
 
-    return min + (int)(mtwist_drand(rl_twister_state) * (max - min + 1));
+    return min + mtwist_drand(rl_twister_state) * (max - min + 1);
 }
 #endif
